+++ src/commands/lock.js
@@ -7,22 +7,20 @@ const _friendlyOS = require("../functions/friendlyOS");
 const _fatalError = require("../functions/fatalError");
 
 const Errors = require("../classes/Errors");
+const InfoMessages = require("../classes/InfoMessages");
 
 const lock = (...args) => {
   try {
     lockSystem();
-
-    console.log(chalk.green(`${GLOBAL_NAME} has successfully locked ${_friendlyOS()}!\n`));
+    InfoMessages.success(`${GLOBAL_NAME} has successfully locked ${_friendlyOS()}!\n`);
     return;
   } catch (err) {
     if (err.message.toLowerCase().includes("unsupported os")) {
       Errors.invalidOS("Windows, macOS and/or Linux");
       return;
     } else if (err.message.toLowerCase().includes("no applicable command found")) {
-      console.log(
-        chalk.red(
-          "Locking the OS failed. Please consider installing xdg-screensaver, gnome-screensaver, cinnamon-screensaver, or dm-tool, and try again.\n"
-        )
+      InfoMessages.error(
+        "Locking the OS failed. Please consider installing xdg-screensaver, gnome-screensaver, cinnamon-screensaver, or dm-tool, and try again."
       );
       return;
     }